Move course routes into own file
- Move path logic into same file
- Make nextAndBackPaths the shareable function
