+++ app/utils/journeys.js
@@ -3,7 +3,10 @@ function originalQuery (req) {
   return originalQueryString ? `?${originalQueryString}` : ''
 }
 
-function nextAndBackPaths (paths, currentPath, query) {
+function nextAndBackPaths (paths, req) {
+  var currentPath = req.path
+  var query = originalQuery(req)
+
   var index = paths.indexOf(currentPath)
   var next = paths[index + 1] || ''
   var back = paths[index - 1] || ''
@@ -15,36 +18,4 @@ function nextAndBackPaths (paths, currentPath, query) {
   }
 }
 
-function pickCoursePaths(req) {
-  const applicationId = req.params.applicationId
-  const courseId = req.params.courseId
-
-  var paths = [
-    `/application/${applicationId}`,
-    `/application/${applicationId}/course/${courseId}/found`,
-    `/application/${applicationId}/course/${courseId}/provider`,
-    `/application/${applicationId}/course/${courseId}/pick`,
-    `/application/${applicationId}/course/${courseId}/create`,
-    `/application/${applicationId}`
-  ]
-
-  return nextAndBackPaths(paths, req.path, originalQuery(req))
-}
-
-function findCoursePaths(req) {
-  const applicationId = req.params.applicationId
-  const courseId = req.params.courseId
-
-  var paths = [
-    `/application/${applicationId}`,
-    `/application/${applicationId}/course/${courseId}/found`,
-    `/application/${applicationId}/course/${courseId}/find`
-  ]
-
-  return nextAndBackPaths(paths, req.path, originalQuery(req))
-}
-
-module.exports = {
-  pickCoursePaths,
-  findCoursePaths
-}
+module.exports = { nextAndBackPaths }
